Use js to toggle bootstrap hamburger menu to avoid line wrapping the menu items.
Static css width breakpoints as implemented in bootstrap-4 ( or 3) can not handle a dynamic number of menu items,
number of items characters or run time applied style sheets.

As a way to avoid the the menu items being wrapped across multiple lines the navbar component sets up a event listener to trigger a function on window resize ( and startup).

If the menu height is larger then the expected menu height the bootstrap hamburger is shown by removing the
bootstrap 'navbar-expand-*' class.
The dynamically found breakpoint is then stored and used as threshold to remove the hamburger menu if the screen size increases beyond the calculated threshold.

- Calc menu height a startup ( use to watch for line wrapping).
- Custom ( to reduce bundle size) debounce function to avoid swapping the processor during resize.
- Toggle hamburger menu dynamicly based on line-wrapping and storing dynamic breakpoint.