fix(menu): Add top menu image to navbar
Copy paste issue, the image above navbar was forgotten when the code moved. Added a test for the
logo above menu bar, to make sure the issue won't occur again since we had many issues with this
feature. Also added a max width to fix another issue that was already present in the old menu as
well.

molgenis/molgenis#8422